LayoutTests/http/tests/navigation/ping-attribute/anchor-cross-origin-from-https-UpgradeMixedContent-expected.txt

CONSOLE MESSAGE: [blocked] The page at https://127.0.0.1:8443/navigation/ping-attribute/anchor-cross-origin-from-https.html requested insecure content from http://localhost:8000/navigation/resources/save-ping.py?test=/navigation/ping-attribute/anchor-cross-origin-from-https.html. This content was blocked and must be served over HTTPS.

Ping not sent
